Removed RealEngine in favor of GhostEngine
