+++ crates/lib3h/src/engine/mod.rs
@@ -1,24 +1,23 @@
+pub mod engine_actor;
 pub mod ghost_engine;
 pub mod ghost_engine_wrapper;
 mod network_layer;
 pub mod p2p_protocol;
-pub mod real_engine;
 mod space_layer;
-
-use std::collections::{HashMap, HashSet, VecDeque};
 
 use crate::{
     dht::dht_protocol::*,
+    error::*,
     gateway::{protocol::*, P2pGateway},
     track::Tracker,
     transport::TransportMultiplex,
 };
 use detach::prelude::*;
 use lib3h_crypto_api::{Buffer, CryptoSystem};
-use lib3h_protocol::{
-    protocol_client::Lib3hClientProtocol, protocol_server::Lib3hServerProtocol, Address,
-};
+use lib3h_ghost_actor::prelude::*;
+use lib3h_protocol::{protocol::*, Address};
 use serde::{ser::SerializeSeq, Deserialize, Deserializer, Serializer};
+use std::collections::{HashMap, HashSet};
 use url::Url;
 
 /// Identifier of a source chain: SpaceAddress+AgentId
@@ -88,40 +87,100 @@ pub struct TransportKeys {
     /// The TransportId secret key
     pub transport_secret_key: Box<dyn Buffer>,
 }
+impl TransportKeys {
+    pub fn new(crypto: &dyn CryptoSystem) -> Lib3hResult<Self> {
+        let hcm0 = hcid::HcidEncoding::with_kind("hcm0")?;
+        let mut public_key: Box<dyn Buffer> = Box::new(vec![0; crypto.sign_public_key_bytes()]);
+        let mut secret_key = crypto.buf_new_secure(crypto.sign_secret_key_bytes());
+        crypto.sign_keypair(&mut public_key, &mut secret_key)?;
+        Ok(Self {
+            transport_id: hcm0.encode(&public_key)?,
+            transport_public_key: public_key,
+            transport_secret_key: secret_key,
+        })
+    }
+}
 
-/// Lib3h's 'real mode' as a NetworkEngine
-pub struct RealEngine {
+pub struct GhostEngine<'engine> {
     /// Identifier
     name: String,
     /// Config settings
     config: RealEngineConfig,
-    /// FIFO of Lib3hClientProtocol messages received from Core
-    inbox: VecDeque<Lib3hClientProtocol>,
-    /// Factory for building the DHTs used by the gateways
+    /// Factory for building DHT's of type D
     dht_factory: DhtFactory,
     /// Tracking request_id's sent to core
     request_track: Tracker<RealEngineTrackerData>,
-
-    /// holds our network gateway and allows connecting routes to space gateways
-    multiplexer: Detach<GatewayParentWrapper<RealEngine, TransportMultiplex<P2pGateway>>>,
-
-    /// Cached this_peer of the network_gateway
+    /// P2p gateway for the network layer
+    multiplexer: Detach<GatewayParentWrapper<GhostEngine<'engine>, TransportMultiplex<P2pGateway>>>,
+    /// Cached this_peer of the multiplexer
     this_net_peer: PeerData,
 
     /// Store active connections?
     network_connections: HashSet<Url>,
-
     /// Map of P2p gateway per Space+Agent
-    space_gateway_map: HashMap<ChainId, Detach<GatewayParentWrapper<RealEngine, P2pGateway>>>,
+    space_gateway_map:
+        HashMap<ChainId, Detach<GatewayParentWrapper<GhostEngine<'engine>, P2pGateway>>>,
     #[allow(dead_code)]
     /// crypto system to use
     crypto: Box<dyn CryptoSystem>,
     #[allow(dead_code)]
     /// transport_id data, public/private keys, etc
     transport_keys: TransportKeys,
-    /// debug: count number of calls to process()
-    process_count: u64,
-    /// dht ghost user_data
-    /// temp HACK. Waiting for gateway actor
-    temp_outbox: Vec<Lib3hServerProtocol>,
+
+    client_endpoint: Option<
+        GhostEndpoint<
+            ClientToLib3h,
+            ClientToLib3hResponse,
+            Lib3hToClient,
+            Lib3hToClientResponse,
+            Lib3hError,
+        >,
+    >,
+    lib3h_endpoint: Detach<
+        GhostContextEndpoint<
+            GhostEngine<'engine>,
+            Lib3hToClient,
+            Lib3hToClientResponse,
+            ClientToLib3h,
+            ClientToLib3hResponse,
+            Lib3hError,
+        >,
+    >,
 }
+
+///// Lib3h's 'real mode' as a NetworkEngine
+//pub struct RealEngine {
+//    /// Identifier
+//    name: String,
+//    /// Config settings
+//    config: RealEngineConfig,
+//    /// FIFO of Lib3hClientProtocol messages received from Core
+//    inbox: VecDeque<Lib3hClientProtocol>,
+//    /// Factory for building the DHTs used by the gateways
+//    dht_factory: DhtFactory,
+//    /// Tracking request_id's sent to core
+//    request_track: Tracker<RealEngineTrackerData>,
+//
+//    /// holds our network gateway and allows connecting routes to space gateways
+//    multiplexer: Detach<GatewayParentWrapper<RealEngine, TransportMultiplex<P2pGateway>>>,
+//
+//    /// Cached this_peer of the network_gateway
+//    this_net_peer: PeerData,
+//
+//    /// Store active connections?
+//    network_connections: HashSet<Url>,
+//
+//    /// Map of P2p gateway per Space+Agent
+//    space_gateway_map: HashMap<ChainId, Detach<GatewayParentWrapper<RealEngine, P2pGateway>>>,
+//    #[allow(dead_code)]
+//    /// crypto system to use
+//    crypto: Box<dyn CryptoSystem>,
+//    #[allow(dead_code)]
+//    /// transport_id data, public/private keys, etc
+//    transport_keys: TransportKeys,
+//    /// debug: count number of calls to process()
+//    process_count: u64,
+//    /// dht ghost user_data
+//    /// temp HACK. Waiting for gateway actor
+//    temp_outbox: Vec<Lib3hServerProtocol>,
+//}
