+++ crates/lib3h/src/engine/space_layer.rs
@@ -3,20 +3,19 @@
 use super::RealEngineTrackerData;
 use crate::{
     dht::dht_protocol::*,
-    engine::{p2p_protocol::SpaceAddress, real_engine::handle_gossipTo, ChainId, RealEngine},
+    engine::{ghost_engine::handle_gossipTo, p2p_protocol::SpaceAddress, ChainId, GhostEngine},
+    error::*,
     gateway::{protocol::*, P2pGateway},
 };
 use detach::prelude::*;
 use lib3h_ghost_actor::prelude::*;
-use lib3h_protocol::{
-    data_types::*, error::Lib3hProtocolResult, protocol_server::Lib3hServerProtocol,
-};
+use lib3h_protocol::{data_types::*, protocol::*, DidWork};
 use lib3h_tracing::Lib3hSpan;
 use std::collections::HashMap;
 
 /// Space layer related private methods
 /// Engine does not process a space gateway's Transport because it is shared with the network layer
-impl RealEngine {
+impl<'engine> GhostEngine<'engine> {
     /// Return list of space+this_peer for all currently joined Spaces
     pub fn get_all_spaces(&mut self) -> Vec<(SpaceAddress, PeerData)> {
         let mut result = Vec::new();
@@ -36,7 +35,7 @@ impl RealEngine {
     pub fn get_first_space_mut(
         &mut self,
         space_address: &str,
-    ) -> Option<&mut GatewayParentWrapper<RealEngine, P2pGateway>> {
+    ) -> Option<&mut GatewayParentWrapper<GhostEngine<'engine>, P2pGateway>> {
         for (chainId, space_gateway) in self.space_gateway_map.iter_mut() {
             let current_space_address: String = chainId.0.clone().into();
             if current_space_address == space_address {
@@ -47,15 +46,12 @@ impl RealEngine {
     }
 
     /// Process all space gateways
-    pub(crate) fn process_space_gateways(
-        &mut self,
-    ) -> Lib3hProtocolResult<Vec<Lib3hServerProtocol>> {
+    pub(crate) fn process_space_gateways(&mut self) -> Lib3hResult<DidWork> {
         // Process all space gateways and collect requests
-        let mut outbox = Vec::new();
         let mut space_outbox_map = HashMap::new();
         let mut space_gateway_map: HashMap<
             ChainId,
-            Detach<GatewayParentWrapper<RealEngine, P2pGateway>>,
+            Detach<GatewayParentWrapper<GhostEngine<'engine>, P2pGateway>>,
         > = self.space_gateway_map.drain().collect();
         for (chain_id, mut space_gateway) in space_gateway_map.drain() {
             detach_run!(space_gateway, |g| g.process(self)).unwrap(); // FIXME unwrap
@@ -70,12 +66,11 @@ impl RealEngine {
         for (chain_id, request_list) in space_outbox_map {
             for mut request in request_list {
                 let payload = request.take_message().expect("exists");
-                let mut output = self.handle_space_request(&chain_id, payload)?;
-                outbox.append(&mut output);
+                self.handle_space_request(&chain_id, payload)?;
             }
         }
         // Done
-        Ok(outbox)
+        Ok(true /* fixme */)
     }
 
     /// Handle a GatewayRequestToParent sent to us by one of our space gateway
@@ -83,12 +78,11 @@ impl RealEngine {
         &mut self,
         chain_id: &ChainId,
         request: GatewayRequestToParent,
-    ) -> Lib3hProtocolResult<Vec<Lib3hServerProtocol>> {
+    ) -> Lib3hResult<DidWork> {
         debug!(
             "{} << handle_space_request: [{:?}] - {:?}",
             self.name, chain_id, request,
         );
-        let mut outbox = Vec::new();
         let space_gateway = self
             .space_gateway_map
             .get_mut(chain_id)
@@ -99,12 +93,8 @@ impl RealEngine {
             GatewayRequestToParent::Dht(dht_request) => {
                 match dht_request {
                     DhtRequestToParent::GossipTo(gossip_data) => {
-                        handle_gossipTo(
-                            &chain_id.0.to_string(),
-                            space_gateway.as_mut(),
-                            gossip_data,
-                        )
-                        .expect("Failed to gossip with space_gateway");
+                        handle_gossipTo(&chain_id.0.to_string(), space_gateway, gossip_data)
+                            .expect("Failed to gossip with space_gateway");
                     }
                     DhtRequestToParent::GossipUnreliablyTo(_data) => {
                         // n/a - should have been handled by gateway
@@ -141,7 +131,12 @@ impl RealEngine {
                                 &lib3h_msg.request_id,
                                 Some(RealEngineTrackerData::HoldEntryRequested),
                             );
-                            outbox.push(Lib3hServerProtocol::HandleStoreEntryAspect(lib3h_msg))
+                            self.lib3h_endpoint
+                                .publish(
+                                    Lib3hSpan::todo(),
+                                    Lib3hToClient::HandleStoreEntryAspect(lib3h_msg),
+                                )
+                                .unwrap(); // FIXME unwrap
                         }
                     }
                     DhtRequestToParent::EntryPruned(_address) => {
@@ -156,7 +151,9 @@ impl RealEngine {
                             provider_agent_id: chain_id.1.clone(),
                             aspect_address_list: None,
                         };
-                        outbox.push(Lib3hServerProtocol::HandleFetchEntry(msg_data))
+                        self.lib3h_endpoint
+                            .publish(Lib3hSpan::todo(), Lib3hToClient::HandleFetchEntry(msg_data))
+                            .unwrap(); // FIXME unwrap
                     }
                 }
             }
@@ -166,6 +163,6 @@ impl RealEngine {
                 // FIXME
             }
         }
-        Ok(outbox)
+        Ok(true /* fixme */)
     }
 }
