+++ crates/lib3h/tests/engine_test.rs
@@ -12,7 +12,7 @@ use predicates::prelude::*;
 
 use lib3h::{
     dht::mirror_dht::MirrorDht,
-    engine::{RealEngine, RealEngineConfig},
+    engine::{GhostEngine, RealEngineConfig},
 };
 use lib3h_protocol::{
     data_types::*, network_engine::NetworkEngine, protocol_client::Lib3hClientProtocol,
@@ -60,7 +60,7 @@ fn enable_logging_for_test(enable: bool) {
 // Engine Setup
 //--------------------------------------------------------------------------------------------------
 
-fn basic_setup_mock_bootstrap(name: &str, bs: Option<Vec<Url>>) -> RealEngine {
+fn basic_setup_mock_bootstrap(name: &str, bs: Option<Vec<Url>>) -> GhostEngine {
     let bootstrap_nodes = match bs {
         Some(s) => s,
         None => vec![],
@@ -76,7 +76,7 @@ fn basic_setup_mock_bootstrap(name: &str, bs: Option<Vec<Url>>) -> RealEngine {
         dht_timeout_threshold: 1000,
         dht_custom_config: vec![],
     };
-    let engine = RealEngine::new_mock(
+    let engine = GhostEngine::new_mock(
         Box::new(SodiumCryptoSystem::new()),
         config,
         name.into(),
@@ -91,7 +91,7 @@ fn basic_setup_mock_bootstrap(name: &str, bs: Option<Vec<Url>>) -> RealEngine {
     engine
 }
 
-fn basic_setup_mock(name: &str) -> RealEngine {
+fn basic_setup_mock(name: &str) -> GhostEngine {
     basic_setup_mock_bootstrap(name, None)
 }
